+++ packages/ente-ui/src/helpers/isSentryDsn.spec.ts
@@ -1,0 +1,20 @@
+import { isSentryDsn } from "./isSentryDsn";
+
+describe("isSentryDsn", () => {
+  const valid = [
+    "https://23bnjkadsk:djaskl@simonknott.de/321",
+    "https://23bnjkadsk:djaskl@sentry.io/321"
+  ];
+  valid.forEach(v => {
+    test(v, () => {
+      expect(isSentryDsn(v)).toBe(true);
+    });
+  });
+
+  const invalid = ["<nil>", "", "invalid"];
+  invalid.forEach(v => {
+    test(v, () => {
+      expect(isSentryDsn(v)).toBe(false);
+    });
+  });
+});