+++ packages/ente-ui/src/helpers/isSentryDsn.ts
@@ -1,0 +1,2 @@
+export const isSentryDsn = (s: string) =>
+  /https:\/\/\w+:\w+@(\w+\.)+\w+\/\d+/.test(s);